# iam-management-web/src/main/resources/openapi/iam.yaml
openapi: 3.0.3
info:
  title: Generic IAM Management API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /iam/admin/templates/{versionId}:
    put:
      tags: [Administration]
      operationId: savePermissionTemplateVersion
      parameters:
        - name: versionId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PermissionTemplateVersionRequest'}
      responses:
        '204': {description: Template version saved}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
  /iam/admin/users:
    get:
      tags: [Administration]
      operationId: listUsers
      parameters:
        - name: afterUserId
          in: query
          schema: {type: integer, format: int64, minimum: 0, default: 0}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        '200':
          description: Cursor-paginated users
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserListResponse'}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
  /iam/admin/users/{userId}:
    put:
      tags: [Administration]
      operationId: saveUser
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserRequest'}
      responses:
        '204': {description: User saved}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
  /iam/admin/users/{userId}/identities:
    get:
      tags: [Administration]
      operationId: listUserIdentities
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        '200':
          description: Identities bound to the user
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/IdentityResponse'}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
  /iam/admin/identities/{identityId}:
    put:
      tags: [Administration]
      operationId: saveIdentity
      parameters:
        - name: identityId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IdentityRequest'}
      responses:
        '204': {description: Identity saved and authorization version incremented}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
  /iam/admin/profiles/{profileId}:
    put:
      tags: [Administration]
      operationId: saveAuthorizationProfile
      parameters:
        - name: profileId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AuthorizationProfileRequest'}
      responses:
        '204': {description: Profile and scopes saved}
        '400': {description: Path and body profile identifiers differ}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
  /iam/admin/profiles/{profileId}/scopes:
    put:
      tags: [Administration]
      operationId: replaceAuthorizationProfileScopes
      parameters:
        - name: profileId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ScopeReplacementRequest'}
      responses:
        '204': {description: Profile scopes replaced and authorization version incremented}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
...
  /iam/admin/sessions/{sessionId}/revoke:
    post:
      tags: [Administration]
      operationId: forceRevokeSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RevokeRequest'}
      responses:
        '204': {description: Session forcibly revoked}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
        '404': {description: Active session not found}
  /iam/admin/users/{userId}/sessions/revoke:
    post:
      tags: [Administration]
      operationId: forceRevokeUserSessions
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RevokeRequest'}
      responses:
        '204': {description: All active sessions for the user forcibly revoked}
        '401': {description: Authentication required}
        '403': {description: Administration permission denied}
components:
  schemas:
...
    RevokeRequest:
      type: object
      properties:
        reason: {type: string, default: USER_REQUEST}
...
    UserRequest:
      type: object
      required: [username, userType, enabled]
      properties:
        username: {type: string, minLength: 1, maxLength: 191}
        userType: {type: string, minLength: 1, maxLength: 64}
        enabled: {type: boolean}
    UserResponse:
      type: object
      required: [userId, username, userType, enabled, authorizationVersion]
      properties:
        userId: {type: integer, format: int64}
        username: {type: string}
        userType: {type: string}
        enabled: {type: boolean}
        authorizationVersion: {type: integer, format: int64}
    UserListResponse:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/UserResponse'}}
        nextAfterUserId: {type: integer, format: int64, nullable: true}
    IdentityRequest:
      type: object
      required: [userId, identityKey, domain, enabled]
      properties:
        userId: {type: integer, format: int64}
        identityKey: {type: string, minLength: 1, maxLength: 191}
        domain: {type: string, minLength: 1, maxLength: 64}
        enabled: {type: boolean}
    IdentityResponse:
      type: object
      required: [identityId, userId, identityKey, domain, enabled]
      properties:
        identityId: {type: string}
        userId: {type: integer, format: int64}
        identityKey: {type: string}
        domain: {type: string}
        enabled: {type: boolean}
    ResourceScope:
      type: object
      required: [scopeType, scopeRefId, accessMode]
      properties:
        scopeType: {type: string}
        scopeRefId: {type: string}
        accessMode: {type: string, enum: [READ, WRITE]}
    ScopeReplacementRequest:
      type: object
      required: [scopes]
      properties:
        scopes:
          type: array
          items: {$ref: '#/components/schemas/ResourceScope'}
    AuthorizationProfileResponse:
      type: object
      required: [profileId, userId, profileName, templateVersionId, clientTypes, enabled, revoked, scopes]
      properties:
        profileId: {type: integer, format: int64}
        userId: {type: integer, format: int64}
        profileName: {type: string}
        templateVersionId: {type: integer, format: int64}
        clientTypes: {type: array, items: {type: string}}
        enabled: {type: boolean}
        revoked: {type: boolean}
        validFrom: {type: string, format: date-time, nullable: true}
        validUntil: {type: string, format: date-time, nullable: true}
        scopes: {type: array, items: {$ref: '#/components/schemas/ResourceScope'}}
    AuthorizationProfileRequest:
      allOf:
        - {$ref: '#/components/schemas/AuthorizationProfileResponse'}
    PermissionTemplateVersionRequest:
      type: object
      required: [templateId, versionNumber, status, permissions]
      properties:
        templateId: {type: integer, format: int64}
        versionNumber: {type: integer, minimum: 1}
        status: {type: string, enum: [DRAFT, PUBLISHED]}
        permissions: {type: array, uniqueItems: true, items: {type: string}}
...